requirements.txt
# Python 3.11 or 3.12 required — pandas does not support Python 3.13+
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
anthropic==0.25.0
python-dotenv==1.0.1
schedule==1.2.1
requests==2.31.0
tabulate==0.9.0
colorlog==6.8.2
pytest==8.1.1
pytest-asyncio==0.23.6
alpaca-trade-api==3.0.2
